BWS-126: * change helpers to buttons
