+++ frontend/src/components/slc/components/functions-table/helpers/cells/actions-row-cell/actions-row-cell.tsx
@@ -1,15 +1,27 @@
-import { DeleteRowCell } from './delete-row-cell/delete-row-cell';
-import { EditRowCell } from './edit-row-cell/edit-row-cell';
+import deleteIcon from 'assets/img/delete-icon.svg';
+import { Link } from 'react-router-dom';
+import { AppRoute } from 'common/enums/enums';
 import styles from './styles.module.scss';
 
 const ActionsRowCell = (
   id: string,
   onFunctionDelete: (id: string) => void,
 ): JSX.Element => {
+  const handleFunctionDelete = (): void => {
+    onFunctionDelete(id);
+  };
+
   return (
     <div className={styles.actionsWrapper}>
-      {DeleteRowCell(id, onFunctionDelete)}
-      {EditRowCell(id)}
+      <button className={styles.button} onClick={handleFunctionDelete}>
+        <img src={deleteIcon} alt="Delete" />
+      </button>
+      <Link
+        className={styles.redirect}
+        to={`${AppRoute.SLC_CONFIGURATE_FUNCTION}/${id}`}
+      >
+        Edit
+      </Link>
     </div>
   );
 };
