--- frontend/src/components/slc/components/functions-table/helpers/cells/actions-row-cell/delete-row-cell/delete-row-cell.tsx
@@ -1,19 +1,0 @@
-import deleteIcon from 'assets/img/delete-icon.svg';
-import styles from './styles.module.scss';
-
-const DeleteRowCell = (
-  id: string,
-  onFunctionDelete: (id: string) => void,
-): JSX.Element => {
-  const handleFunctionDelete = (): void => {
-    onFunctionDelete(id);
-  };
-
-  return (
-    <button className={styles.button} onClick={handleFunctionDelete}>
-      <img src={deleteIcon} alt="Delete" />
-    </button>
-  );
-};
-
-export { DeleteRowCell };
