--- frontend/src/components/slc/components/functions-table/helpers/cells/actions-row-cell/delete-row-cell/styles.module.scss
@@ -1,5 +1,0 @@
-.button {
-  background: inherit;
-  border: none;
-  cursor: pointer;
-}
